todoapp/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==37.0.2
Django==4.0.4
django-environ==0.8.1
mysqlclient==2.1.0
pycparser==2.21
PyJWT==2.4.0
sqlparse==0.4.2
